Refactor the GetStrippedOffsets
This one handles ... more elegantly all cases where the execution
segmetn sequence isn't full or missing at all. This is mostly done
through the fact that the algorithm works even if we don't fill the
sequence as long as we don't care about whether 0:1/3 and 1/3:2/3 get
the same offsets when there is no sequence provided. If we care about
that we need to fill the sequence which will be more involved but not by
much :D.

The other thing is that the construction of the numerator, steps and so
on arrays is separate from the actual algorithm of calculating the
offsets which at least makes silence golangci-lint. I would argue also
makes it easier to read.